Started creation for server load balancing
When the react app loads it will conect to a master php server, which will query for a slave server from the db. this server will be sent back to the react app. 
Also need to think about how to setup multiple react servers. Maybe a php front page that redirects to an available react server.